// ui/src/types/TokenModel.ts
export interface TokenModel {
  name: string;
  symbol: string;
  price: number;
  icon: string;
  balance: number;
  address: string;
  // isNative: any;
  // isToken: any;
  // list: any;
  // tokenInfo: any;
  // chainId: number;
  // equals: any,
  // sortsBefore: any,
  // wrapped: any,
  // decimals: number,
  // NativeCurrency: any,
  // Currency: any
}